# perf(mobile): reduzir TBT/LCP na home e /servicos (split & serverize)