Home: dynamic(ssr:false) em Testimonials, Contact e HomeCTAs. /servicos: grid como server component (sem hidratacao) + prefetch=false. Sem mudancas em rotas/UTMs/API. LHCI mobile (2 runs, devtools) medianas: '/' LCP~2841ms TBT~245ms; '/servicos' LCP~2940ms TBT~249ms. Cumpre budgets.